refactor: move the model creation methods
Move the methods in the KitsuLibrary into their own file to handle creating models
